refactor: reworked project structure for better fsd implementation
