+++ src/entities/todo/model/index.ts
@@ -1,2 +1,3 @@
 export type { Todo, TodoStore, TodoDone, TodoNotDone } from './model';
 export { TodoSchema } from './schema';
+export { useTodoStore, type TodoFilter} from './store';
