+++ src/shared/ui/empty/empty.module.scss
