--- src/shared/ui/empty/ui/index.ts
@@ -1,1 +1,0 @@
-export { Empty } from './empty';
