--- src/shared/ui/loading-screen/ui/index.ts
@@ -1,1 +1,0 @@
-export { LoadingScreen } from './loading-screen';
