--- src/shared/ui/todo-dialog/ui/index.ts
@@ -1,1 +1,0 @@
-export { TodoDialog } from './todo-dialog';
