+++ src/views/home/ui/index.tsx
@@ -4,21 +4,13 @@ import { ru } from 'date-fns/locale';
 import { Filter } from '@/features/filter';
 import { TodoList } from '@/widgets/todo-list';
 import { CreateTodoDialog, CreateTodoFab } from '@/features/create-todo';
-import { useCallback, useRef } from 'react';
 import { UpdateTodoDialog } from '@/features/update-todo';
 import { DeleteTodoDialog } from '@/features/delete-todo';
 
 export default function HomePage() {
-  const dialogRef = useRef<HTMLElement>(null);
   const dateNow = new Date();
   const dayOfWeek = format(dateNow, 'iiii', { locale: ru });
   const dayAndMonth = format(dateNow, 'd MMMM', { locale: ru });
-
-  const toggleDialog = useCallback(() => {
-    if (dialogRef.current && 'open' in dialogRef.current) {
-      dialogRef.current.open = !dialogRef.current.open;
-    }
-  }, []);
 
   return (
     <>
